Removed #pragma from OP2Helper.
The #pragma made an unneeded relative path assumption. Setting a proper project dependency and reference is a better way to find the needed library, and also make the API headers easily browsable. Alternatively, the .lib file could be added as a resource to projects that include the OP2Helper API.

git-svn-id: https://svn.outpostuniverse.org:8443/svn/outpost2/LevelsAndMods/trunk/API/OP2Helper@967 7b3b2116-5498-11dd-abe4-fb2d8bf8a5a8