Remove reference semantics from non-const temp
Capturing a temporary by non-const reference is an MSVC compiler
extension of the C++ standard, which may issue a warning with newer
compilers. Using a const reference would be standards compliant, but not
appropriate here since the value is later used in a non-const manner.

Some consideration was given to new C++11 rvalue syntax (&&), though it
seems most new compilers can optimize away the object copy without this,
and the preferred syntax is to just omit reference syntax when calling
factory functions. At any rate, the object is small, so performance
differences would be minimal in any case.

The bigger concern is perhaps keeping the library simple. This library
will be used, and possibly extended, by people without a lot of C++
experience. Making the library simple to understand and use should by
the highest priority.